requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
apispec==6.3.0
asgiref==3.7.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==5.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
fastapi==0.105.0
Flask==3.0.0
flask-apispec==0.11.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
webargs==8.3.0
Werkzeug==3.0.1
